+++ Samples/iOS-Swift/iOS-Swift/ViewControllers/NibViewController.swift
@@ -22,21 +22,8 @@ class NibViewController: UIViewController {
         UIAssert.checkForViewControllerLifeCycle(span, viewController: "NibViewController")
     }
 
-    override func viewWillAppear(_ animated: Bool) {
-        super.viewWillAppear(animated)
-        SentrySDK.reportFullyDisplayed()
-    }
-
     override func viewDidAppear(_ animated: Bool) {
         super.viewDidAppear(animated)
+        SentrySDK.reportFullyDisplayed()
     }
-    
-    override func viewWillLayoutSubviews() {
-        super.viewWillLayoutSubviews()
-    }
-    
-    override func viewDidLayoutSubviews() {
-        super.viewDidLayoutSubviews()
-    }
-    
 }