+++ Tests/SentryTests/Integrations/SentryCrash/SentryCrashReportTests.swift
@@ -1,3 +1,4 @@
+import SentryTestUtils
 import XCTest
 
 class SentryCrashReportTests: XCTestCase {
@@ -32,6 +33,7 @@ class SentryCrashReportTests: XCTestCase {
         super.tearDown()
         
         deleteTestDir()
+        clearTestState()
     }
         
     func testScopeInCrashReport_IsSameAsSerializingIt() {
@@ -76,7 +78,7 @@ class SentryCrashReportTests: XCTestCase {
      * UserInfo is picked up by the crash report when writing a new report.
      */
     private func serializeToCrashReport(scope: Scope) {
-        SentryCrash.sharedInstance().userInfo = scope.serialize()
+        SentryDependencyContainer.sharedInstance().crashReporter.userInfo = scope.serialize()
     }
     
     private func deleteTestDir() {