+++ Tests/SentryTests/SentryCrash/SentryCrashInstallationTests.m
@@ -5,6 +5,7 @@
 #import "SentryCrashMonitor.h"
 #import "SentryCrashMonitor_MachException.h"
 #import "SentryCrashSystemCapabilities.h"
+#import "SentryDependencyContainer.h"
 #import "SentryNSNotificationCenterWrapper.h"
 #import "SentryTests-Swift.h"
 #import <XCTest/XCTest.h>
@@ -33,14 +34,13 @@ NS_ASSUME_NONNULL_BEGIN
 
 @implementation SentryCrashInstallationTests
 
-- (SentryCrashTestInstallation *)getSut
+- (void)tearDown
 {
-    SentryCrashTestInstallation *installation =
-        [[SentryCrashTestInstallation alloc] initForTesting];
-    self.notificationCenter = [[TestNSNotificationCenterWrapper alloc] init];
-    [[SentryCrash sharedInstance] setSentryNSNotificationCenterWrapper:self.notificationCenter];
-    return installation;
+    [super tearDown];
+    [SentryDependencyContainer reset];
 }
+
+#pragma mark - Tests
 
 - (void)testUninstall
 {
@@ -48,7 +48,8 @@ NS_ASSUME_NONNULL_BEGIN
 
     [installation install];
 
-    SentryCrashMonitorType monitorsAfterInstall = [SentryCrash sharedInstance].monitoring;
+    SentryCrashMonitorType monitorsAfterInstall
+        = SentryDependencyContainer.sharedInstance.crashReporter.monitoring;
 
     [installation uninstall];
 
@@ -73,8 +74,8 @@ NS_ASSUME_NONNULL_BEGIN
 
     [installation install];
 
-    SentryCrash *sentryCrash = [SentryCrash sharedInstance];
-    SentryCrashMonitorType monitorsAfterInstall = sentryCrash.monitoring;
+    SentryCrashMonitorType monitorsAfterInstall
+        = SentryDependencyContainer.sharedInstance.crashReporter.monitoring;
     CrashHandlerData *crashHandlerDataAfterInstall = [installation g_crashHandlerData];
 
     // To ensure multiple calls in a row work
@@ -97,14 +98,26 @@ NS_ASSUME_NONNULL_BEGIN
 
 #if SentryCrashCRASH_HAS_UIAPPLICATION
     XCTAssertEqual(55, self.notificationCenter.removeObserverWithNameInvocationsCount);
-#endif
+#endif // SentryCrashCRASH_HAS_UIAPPLICATION
+}
+
+#pragma mark - Private
+
+- (SentryCrashTestInstallation *)getSut
+{
+    SentryCrashTestInstallation *installation =
+        [[SentryCrashTestInstallation alloc] initForTesting];
+    self.notificationCenter = [[TestNSNotificationCenterWrapper alloc] init];
+    [SentryDependencyContainer.sharedInstance.crashReporter
+        setSentryNSNotificationCenterWrapper:self.notificationCenter];
+    return installation;
 }
 
 - (void)assertReinstalled:(SentryCrashTestInstallation *)installation
             monitorsAfterInstall:(SentryCrashMonitorType)monitorsAfterInstall
     crashHandlerDataAfterInstall:(CrashHandlerData *)crashHandlerDataAfterInstall
 {
-    SentryCrash *sentryCrash = [SentryCrash sharedInstance];
+    SentryCrash *sentryCrash = SentryDependencyContainer.sharedInstance.crashReporter;
     XCTAssertNotEqual(NULL, [installation g_crashHandlerData]);
     XCTAssertEqual(monitorsAfterInstall, sentryCrash.monitoring);
     XCTAssertEqual(monitorsAfterInstall, sentrycrashcm_getActiveMonitors());
@@ -119,7 +132,7 @@ NS_ASSUME_NONNULL_BEGIN
 - (void)assertUninstalled:(SentryCrashTestInstallation *)installation
      monitorsAfterInstall:(SentryCrashMonitorType)monitorsAfterInstall
 {
-    SentryCrash *sentryCrash = [SentryCrash sharedInstance];
+    SentryCrash *sentryCrash = SentryDependencyContainer.sharedInstance.crashReporter;
     XCTAssertEqual(NULL, [installation g_crashHandlerData]);
     XCTAssertEqual(SentryCrashMonitorTypeNone, sentryCrash.monitoring);
     XCTAssertEqual(SentryCrashMonitorTypeNone, sentrycrashcm_getActiveMonitors());