+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Unreleased
+
+### Fixes
+
+- Crash when initializing SentryHub manually (#3374)
+
 ## 8.15.0
 
 ### Features
@@ -15,7 +21,6 @@
 - Thread sanitizer data race warnings in ANR tracker, network tracker and span finish (#3303)
 - Stop sending empty thread names (#3361)
 - Work around edge case with a thread info kernel call sometimes returning invalid data, leading to a crash (#3364)
-- Crash when initializing SentryHub manually (#3374)
 - Crashes when trace ID is externally modified or profiler fails to initialize (#3365)
 
 ## 8.14.2