+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # Changelog
 
-## Unreleased
+## 8.15.1
+
+### Fixes
+
+- Crash when initializing SentryHub manually (#3374)
+
+## 8.15.0
 
 ### Features